+++ src/lib/packages/types.ts
@@ -3,9 +3,9 @@ import type Resource from "../resources/resource";
 /**
  * Pairing of a resource key and resource it belongs to.
  */
-export interface ResourceKeyPair {
+export interface ResourceKeyPair<ResourceType extends Resource = Resource> {
   key: ResourceKey;
-  value: Resource;
+  value: ResourceType;
 }
 
 /**