+++ src/lib/resources/simdata/constants.ts
@@ -1,5 +1,0 @@
-export const RELOFFSET_NULL = -0x80000000;
-export const NO_NAME_HASH = 0x811C9DC5; // equal to fnv32('')
-export const HEADER_SIZE = 28; // does not include bytes for padding
-export const TABLE_HEADER_OFFSET = 24;
-export const SUPPORTED_VERSION = 0x101;