+++ src/models.ts
@@ -3,11 +3,13 @@ import XmlResource from "./lib/resources/xml/xml-resource";
 import RawResource from "./lib/resources/raw/raw-resource";
 import SimDataResource from "./lib/resources/simdata/simdata-resource";
 import StringTableResource from "./lib/resources/stbl/stbl-resource";
+import CombinedTuningResource from "./lib/resources/combined-tuning/combined-tuning-resource";
 
 export {
   Package,
   RawResource,
   XmlResource,
   SimDataResource,
-  StringTableResource
+  StringTableResource,
+  CombinedTuningResource
 }